+++ tests/functions/test_encode_categorical.py
@@ -2,11 +2,8 @@ import numpy as np
 import pandas as pd
 import pytest
 from hypothesis import given
-from pandas.api.types import CategoricalDtype
 from pandas.testing import assert_frame_equal
 
-from janitor import As_Categorical
-from janitor.errors import JanitorError
 from janitor.testing_utils.strategies import (
     categoricaldf_strategy,
     df_strategy,
@@ -24,11 +21,11 @@ def test_encode_categorical(df):
 @given(df=df_strategy())
 def test_encode_categorical_missing_column(df):
     """
-    Raise JanitorError for missing columns
+    Raise ValueError for missing columns
     when only one arguments is provided to
     `column_names`.
     """
-    with pytest.raises(JanitorError):
+    with pytest.raises(ValueError):
         df.encode_categorical("aloha")
 
 
@@ -36,11 +33,11 @@ def test_encode_categorical_missing_column(df):
 @given(df=df_strategy())
 def test_encode_categorical_missing_columns(df):
     """
-    Raise JanitorError for missing columns
+    Raise ValueError for missing columns
     when the number of arguments to `column_names`
     is more than one.
     """
-    with pytest.raises(JanitorError):
+    with pytest.raises(ValueError):
         df.encode_categorical(["animals@#$%^", "cities", "aloha"])
 
 
@@ -48,10 +45,10 @@ def test_encode_categorical_missing_columns(df):
 @given(df=df_strategy())
 def test_encode_categorical_invalid_input(df):
     """
-    Raise JanitorError for wrong input type
+    Raise ValueError for wrong input type
     for `column_names`.
     """
-    with pytest.raises(JanitorError):
+    with pytest.raises(ValueError):
         df.encode_categorical(1)
 
 
@@ -59,75 +56,100 @@ def test_encode_categorical_invalid_input(df):
 @given(df=df_strategy())
 def test_encode_categorical_invalid_input_2(df):
     """
-    Raise JanitorError for wrong input type
+    Raise TypeError for wrong input type
     for `column_names`.
     """
-    with pytest.raises(JanitorError):
+    with pytest.raises(TypeError):
         df.encode_categorical({"names"})
 
 
-@pytest.fixture
-def df_categorical():
-    return pd.DataFrame(
-        {
-            "col1": [2, 1, 3, 1],
-            "col2": ["a", "b", "c", "d"],
-            "col3": pd.date_range("1/1/2020", periods=4),
-        }
-    )
-
-
-def test_encode_categorical_multiple_column_names(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_encode_categorical_multiple_column_names(df):
     """
     Test output when more than one column is provided
     to `column_names`.
     """
-    result = df_categorical.astype({"col1": "category", "col2": "category"})
+    result = df.astype({"a": "category", "cities": "category"})
     assert_frame_equal(
-        df_categorical.encode_categorical(column_names=["col1", "col2"]),
+        df.encode_categorical(column_names=["a", "cities"]),
         result,
     )
 
 
-def test_both_column_names_kwargs(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_both_column_names_kwargs(df):
     """
     Raise Error if both `column_names`
     and kwargs are provided.
     """
     with pytest.raises(ValueError):
-        df_categorical.encode_categorical(
-            column_names=["col1", "col2"], col1=(None, "sort")
+        df.encode_categorical(
+            column_names=["a", "cities"], Bell__Chart=(None, "sort")
         )
 
 
-def test_check_presence_column_names_in_kwargs(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_check_presence_column_names_in_kwargs(df):
     """
     Raise ValueError if column names in `kwargs`
     do not exist in the dataframe.
     """
     with pytest.raises(ValueError):
-        df_categorical.encode_categorical(col_1=(None, "sort"))
+        df.encode_categorical(col_1=(None, "sort"))
 
 
-def test_check_type_tuple_in_kwargs(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_check_type_tuple_in_kwargs(df):
     """
     Raise TypeError if the categories, order pairing
     in `kwargs` is not a tuple.
     """
     with pytest.raises(TypeError):
-        df_categorical.encode_categorical(col1=[None, "sort"])
+        df.encode_categorical(a=[None, "sort"])
 
 
-def test_categories_type_in_kwargs(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_tuple_length_in_kwargs(df):
+    """
+    Raise ValueError if the length of the tuple
+    in kwargs is not equal to 2.
+    """
+    with pytest.raises(ValueError):
+        df.encode_categorical(a=(None, None, 2))
+
+
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_categories_type_in_kwargs(df):
     """
     Raise TypeError if the wrong argument is supplied to
     the `categories` parameter in kwargs.
     """
     with pytest.raises(TypeError):
-        df_categorical.encode_categorical(col1=("category", None))
+        df.encode_categorical(a=("category", None))
 
 
-def test_categories_ndim_gt_1_in_kwargs(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_categories_ndim_array_gt_1_in_kwargs(df):
+    """
+    Raise ValueError if the argument supplied to
+    the `categories` parameter in kwargs is not
+    1-D array-like.
+    """
+    arrays = [[1, 1, 2, 2], ["red", "blue", "red", "blue"]]
+    with pytest.raises(ValueError):
+        df.encode_categorical(a=(arrays, None))
+
+
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_categories_ndim_MultiIndex_gt_1_in_kwargs(df):
     """
     Raise ValueError if the argument supplied to
     the `categories` parameter in kwargs is not
@@ -136,341 +158,179 @@ def test_categories_ndim_gt_1_in_kwargs(df_categorical):
     arrays = [[1, 1, 2, 2], ["red", "blue", "red", "blue"]]
     arrays = pd.MultiIndex.from_arrays(arrays, names=("number", "color"))
     with pytest.raises(ValueError):
-        df_categorical.encode_categorical(col1=(arrays, None))
+        df.encode_categorical(a=(arrays, None))
 
 
-def test_categories_null_in_categories(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_categories_ndim_DataFrame_gt_1_in_kwargs(df):
+    """
+    Raise ValueError if the argument supplied to
+    the `categories` parameter in kwargs is not
+    1-D array-like.
+    """
+    arrays = {"name": [1, 1, 2, 2], "number": ["red", "blue", "red", "blue"]}
+    arrays = pd.DataFrame(arrays)
+    with pytest.raises(ValueError):
+        df.encode_categorical(a=(arrays, None))
+
+
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_categories_null_in_categories(df):
     """
     Raise ValueError if there are nulls in the `categories`.
     """
     with pytest.raises(ValueError):
-        df_categorical.encode_categorical(col1=([None, 2, 3], None))
+        df.encode_categorical(a=([None, 2, 3], None))
 
 
-def test_order_type_in_kwargs(df_categorical):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_non_unique_cat(df):
+    """Raise ValueError if `categories` is not unique."""
+    with pytest.raises(ValueError):
+        df.encode_categorical(a=([1, 2, 3, 3], "sort"))
+
+
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_empty_cat(df):
+    """Raise ValueError if `categories` is empty."""
+    with pytest.raises(ValueError):
+        df.encode_categorical(a=([], "sort"))
+
+
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_empty_col(df):
     """
-    Raise TypeError if the wrong argument is supplied to
-    the `order` parameter in kwargs.
-    """
-    with pytest.raises(TypeError):
-        df_categorical.encode_categorical(col1=({1, 2, 3, 3}, {"sort"}))
-
-
-def test_order_wrong_option_in_kwargs(df_categorical):
-    """
-    Raise ValueError if the value supplied to the `order`
-    parameter in kwargs is not one of None, 'sort', or 'appearance'.
+    Raise ValueError if `categories`,
+    and the relevant column is empty, or all nulls.
     """
     with pytest.raises(ValueError):
-        df_categorical.encode_categorical(col1=({1, 2, 3, 3}, "sorted"))
+        df["col1"] = np.nan
+        df.encode_categorical(col1=([1, 2, 3], "sort"))
 
 
-as_categorical_tuple_length = [
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ),
-        (None,),
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ),
-        (None, "sort", "appearance"),
-    ),
-]
-
-
-@pytest.mark.parametrize(
-    "df,as_categorical_tuple", as_categorical_tuple_length
-)
-def test_tuple_length_in_kwargs(df, as_categorical_tuple):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_empty_col_2(df):
     """
-    Raise ValueError if the length of the tuple
-    in kwargs is not equal to 2.
+    Raise ValueError if `categories` is None,
+    and the relevant column is empty, or all nulls.
     """
     with pytest.raises(ValueError):
-        df.encode_categorical(col1=as_categorical_tuple)
+        df["col1"] = np.nan
+        df.encode_categorical(col1=(None, "sort"))
 
 
-test_various_df = [
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ).astype({"col1": "category", "col2": "category"}),
-        {
-            "col1": As_Categorical(categories=None, order=None),
-            "col2": As_Categorical(order=None, categories=None),
-        },
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ).astype({"col1": "category", "col2": "category"}),
-        {"col1": (None, None), "col2": (None, None)},
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ).astype(
-            {
-                "col3": "category",
-                "col1": CategoricalDtype(categories=[1, 2, 3], ordered=True),
-            }
-        ),
-        {"col3": (None, None), "col1": (None, "sort")},
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, np.nan],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2.0, 1, 3, np.nan],
-                "col2": ["a", "b", "c", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ).astype(
-            {
-                "col3": "category",
-                "col1": CategoricalDtype(categories=[1, 2.0, 3], ordered=True),
-            }
-        ),
-        {"col3": (None, None), "col1": (None, "sort")},
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, np.nan],
-                "col2": ["b", "c", "a", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2.0, 1, 3, np.nan],
-                "col2": ["b", "c", "a", "d"],
-                "col3": pd.date_range("1/1/2020", periods=4),
-            }
-        ).astype(
-            {
-                "col2": CategoricalDtype(
-                    categories=["b", "c", "a", "d"], ordered=True
-                ),
-                "col1": CategoricalDtype(categories=[2.0, 1, 3], ordered=True),
-            }
-        ),
-        {"col2": (None, "appearance"), "col1": (None, "appearance")},
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ).astype(
-            {
-                "col1": CategoricalDtype(
-                    categories=[2.0, 1.0, 3.0], ordered=True
-                ),
-                "col2": CategoricalDtype(
-                    categories=["a", "b", "c", "d"], ordered=False
-                ),
-            }
-        ),
-        {
-            "col2": (["a", "b", "c", "d"], None),
-            "col1": ([2.0, 1.0, 3.0], "appearance"),
-        },
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ).astype(
-            {
-                "col1": CategoricalDtype(
-                    categories=[2.0, 1.0, 3.0], ordered=True
-                ),
-                "col2": CategoricalDtype(
-                    categories=["a", "b", "c", "d"], ordered=True
-                ),
-            }
-        ),
-        {
-            "col2": (["a", "b", "c", "d"], "sort"),
-            "col1": ([2.0, 1.0, 3.0], "appearance"),
-        },
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ),
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ).astype(
-            {
-                "col1": CategoricalDtype(
-                    categories=[1.0, 2.0, 3.0], ordered=True
-                ),
-                "col2": CategoricalDtype(
-                    categories=["a", "b", "c", "d"], ordered=True
-                ),
-                "col3": CategoricalDtype(
-                    categories=pd.to_datetime(
-                        [
-                            "2020-01-01",
-                            "2020-01-02",
-                            "2020-01-03",
-                            "2020-01-04",
-                            "2020-01-05",
-                        ]
-                    ),
-                    ordered=True,
-                ),
-            }
-        ),
-        {
-            "col3": (None, "sort"),
-            "col1": (None, "sort"),
-            "col2": (None, "appearance"),
-        },
-    ),
-]
-
-
-@pytest.mark.parametrize("df_in, df_out, kwargs", test_various_df)
-def test_various(df_in, df_out, kwargs):
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_empty_col_3(df):
     """
-    Test output for various combinations.
+    Raise ValueError if `categories` is None,
+    and the relevant column is empty, or all nulls.
     """
-    # result = df_in.encode_categorical(
-    #     **kwargs,
-    # )
-    # assert_frame_equal(result, df_out)
-    result = df_in.encode_categorical(
-        **kwargs,
-    )
-    # directly comparing columns is safe -
-    # if the columns have differing categories
-    # (especially for ordered True) it will fail.
-    # if both categories are unordered, then the
-    # order is not considered.
-    # comparing with assert_frame_equal fails
-    # for unordered categoricals, as internally
-    # the order of the categories are compared
-    # which is irrelevant for unordered categoricals
-    for key, _ in kwargs.items():
-        assert df_out[key].equals(result[key])
+    with pytest.raises(ValueError):
+        df["col1"] = pd.Series([], dtype="object")
+        df.encode_categorical(col1=(None, "appearance"))
 
 
-df_warnings = [
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ),
-        {
-            "col1": ([4, 5, 6], "appearance"),
-            "col2": (["a", "b", "c", "d"], "sort"),
-        },
-    ),
-    (
-        pd.DataFrame(
-            {
-                "col1": [2, 1, 3, 1, np.nan],
-                "col2": ["a", "b", "c", "d", "a"],
-                "col3": pd.date_range("1/1/2020", periods=5),
-            }
-        ),
-        {
-            "col1": ([2.0, 1.0, 3.0], "appearance"),
-            "col2": (["a", "b", "c"], "sort"),
-        },
-    ),
-]
-
-
-@pytest.mark.parametrize("df_in,kwargs", df_warnings)
-def test_warnings(df_in, kwargs):
+@pytest.mark.functions
+@given(df=categoricaldf_strategy())
+def test_warnings(df):
     """
     Test that warnings are raised if `categories` is provided, and
     the categories do not match the unique values in the column, or
     some values in the column are missing in `categories`.
     """
     with pytest.warns(UserWarning):
-        df_in.encode_categorical(**kwargs)
+        df.encode_categorical(
+            numbers=([4, 5, 6], None), names=(["John", "Mark", "Luke"], "sort")
+        )
+
+
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_order_type_in_kwargs(df):
+    """
+    Raise TypeError if the wrong argument is supplied to
+    the `order` parameter in kwargs.
+    """
+    with pytest.raises(TypeError):
+        df.encode_categorical(a=({1, 2, 3, 3}, {"sort"}))
+
+
+@pytest.mark.functions
+@given(df=df_strategy())
+def test_order_wrong_option_in_kwargs(df):
+    """
+    Raise ValueError if the value supplied to the `order`
+    parameter in kwargs is not one of None, 'sort', or 'appearance'.
+    """
+    with pytest.raises(ValueError):
+        df.encode_categorical(a=({1, 2, 3, 3}, "sorted"))
+
+
+# directly comparing columns is safe -
+# if the columns have differing categories
+# (especially for ordered True) it will fail.
+# if both categories are unordered, then the
+# order is not considered.
+# comparing with assert_frame_equal fails
+# for unordered categoricals, as internally
+# the order of the categories are compared
+# which is irrelevant for unordered categoricals
+
+
+@pytest.mark.functions
+@given(df=categoricaldf_strategy())
+def test_all_None(df):
+    """
+    Test output where `categories` and `order` are None.
+    """
+    result = df.encode_categorical(names=(None, None))
+
+    expected = df.astype({"names": "category"})
+    assert expected["names"].equals(result["names"])
+
+
+@pytest.mark.functions
+@given(df=categoricaldf_strategy())
+def test_all_cat_None_1(df):
+    """
+    Test output where `categories` is None.
+    """
+    result = df.encode_categorical(names=(None, "sort"))
+    categories = pd.CategoricalDtype(
+        categories=df.names.factorize(sort=True)[-1], ordered=True
+    )
+    expected = df.astype({"names": categories})
+    assert expected["names"].equals(result["names"])
+
+
+@pytest.mark.functions
+@given(df=categoricaldf_strategy())
+def test_all_cat_None_2(df):
+    """
+    Test output where `categories` is None.
+    """
+    result = df.encode_categorical(names=(None, "appearance"))
+    categories = pd.CategoricalDtype(
+        categories=df.names.factorize(sort=False)[-1], ordered=True
+    )
+    expected = df.astype({"names": categories})
+    assert expected["names"].equals(result["names"])
+
+
+@pytest.mark.functions
+@given(df=categoricaldf_strategy())
+def test_all_cat_not_None(df):
+    """
+    Test output where `categories` is  not None.
+    """
+    result = df.encode_categorical(numbers=(np.array([3, 1, 2]), "appearance"))
+    categories = pd.CategoricalDtype(categories=[3, 1, 2], ordered=True)
+    expected = df.astype({"numbers": categories})
+    assert expected["numbers"].equals(result["numbers"])